Farmer-Agent-backend/requirements.txt
flask
flask-cors
requests
python-dotenv
gradio
google-generativeai
pillow
assemblyai
langdetect